+++ packages/cmd/src/CommandProcessor.ts
@@ -181,10 +181,6 @@ export class CommandProcessor {
         return this.mCommandLine;
     }
 
-    // set commandLine(command: string) {
-    //     this.mCommandLine = command;
-    // }
-
     /**
      * Accessor for the environment variable prefix
      * @readonly
@@ -204,7 +200,6 @@ export class CommandProcessor {
     get promptPhrase(): string {
         return this.mPromptPhrase;
     }
-
 
     /**
      * Accessor for the help generator passed to this instance of the command processor
@@ -370,6 +365,16 @@ export class CommandProcessor {
 
         // Build the response object, base args object, and the entire array of options for this command
         // Assume that the command succeed, it will be marked otherwise under the appropriate failure conditions
+        if (params.arguments.dcd) {
+            // NOTE(Kelosky): we adjust `cwd` and do not restore it, so that multiple simultaneous requests from the same
+            // directory will operate without unexpected chdir taking place.  Multiple simultaneous requests from different
+            // directories may cause unpredictable results
+            process.chdir(params.arguments.dcd as string)
+
+            // reinit config for daemon client directory
+            ImperativeConfig.instance.config = await Config.load(ImperativeConfig.instance.rootCommandName, ImperativeConfig.instance.config.opts);
+            this.mConfig = ImperativeConfig.instance.config;
+        }
         const prepareResponse = this.constructResponseObject(params);
         prepareResponse.succeeded();