+++ CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to the Imperative package will be documented in this file.
 
 - Enhancement: Allow comments to be added to any `.config.json` generated
 
+## `5.0.0-next.202010301408`
+
+- Enhancement: Allow hidden options.
+
 ## `5.0.0-next.202010161240`
 
 - Enhancement:  Allow process exit code to be passed to daemon clients.
@@ -13,6 +17,15 @@ All notable changes to the Imperative package will be documented in this file.
 ## `5.0.0-next.202009251501`
 
 - Enhancement: add support for CLIs that want to run as a persistent process (daemon mode).
+
+## `4.10.0`
+
+- Enhancement: Added an `arrayAllowDuplicate` option to the `ICommandOptionDefinition` interface. By default, the option value is set to `true` and duplicate values are allowed in an array. Specify `false` if you want Imperative to throw an error for duplicate array values. [#437](https://github.com/zowe/imperative/issues/437)
+
+## `4.9.0`
+
+- BugFix: Updated `opener` dependency due to command injection vulnerability on Windows - [GHSL-2020-145](https://securitylab.github.com/advisories/GHSL-2020-145-domenic-opener)
+- Enhancement: Expose `trim` parameter from `wrap-ansi` within `TextUtils.wordWrap()`
 
 ## `4.8.1`